+++ initializers/__init__.py
@@ -1,0 +1,1 @@
+from .jvm import init_jvm